+++ apm-sniffer/apm-sdk-plugin/jdbc-commons/src/main/java/org/apache/skywalking/apm/plugin/jdbc/PSSetterDefinitionOfJDBCInstrumentation.java
@@ -40,11 +40,10 @@ public class PSSetterDefinitionOfJDBCInstrumentation implements InstanceMethodsI
     public ElementMatcher<MethodDescription> getMethodsMatcher() {
         ElementMatcher.Junction<MethodDescription> matcher = none();
 
-        if (JDBCPluginConfig.Plugin.JDBC.TRACE_SQL_PARAMETERS) {
-            final Set<String> setters = ignorable ? PS_IGNORABLE_SETTERS : PS_SETTERS;
-            for (String setter : setters) {
-                matcher = matcher.or(named(setter));
-            }
+        // remove TRACE_SQL_PARAMETERS judgement for dynamic config
+        final Set<String> setters = ignorable ? PS_IGNORABLE_SETTERS : PS_SETTERS;
+        for (String setter : setters) {
+            matcher = matcher.or(named(setter));
         }
 
         return matcher;